Removed controller as integration test, temporarily
